+++ src/components/ShareButton/CopyButton/index.js
@@ -1,12 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {Root, Info, TextInfo} from './CopyButtonLink.styled';
+
 import renderIcon from '../../../utils/renderIcon';
+
+import {Root, Info, TextInfo} from './styled.index';
+
 
 const CopyLinkButton = ({copyTitle, handleCopyLink}) => (
 	<Root onClick={handleCopyLink} data-qa="share-copy">
 		<Info>
-			{renderIcon('copylink')}
+			{renderIcon('copyLink')}
 			<TextInfo>
 				{copyTitle || 'Копировать ссылку'}
 			</TextInfo>
